refactor: move display tests into separate files
